+++ MySuperApp/Class/Databsae/other/EADataModel/修改密码/UpdatepwdupdatepwdModel.h
@@ -14,7 +14,7 @@
 @property (nonatomic, retain) NSString *res;
 @property (nonatomic, retain) NSString *response;
 
-@property (nonatomic, assign) int requestTag;
+@property (nonatomic, assign) NSInteger requestTag;
 @property (retain, nonatomic) NSString *errorMessage;
 
 + (UpdatepwdupdatepwdModel *)modelObjectWithDictionary:(NSDictionary *)dict;
