+++ MySuperApp/Class/Databsae/other/EADataModel/商品列表/ProductlistModel.h
@@ -21,7 +21,7 @@
 @property (nonatomic, assign) double pageCount;
 
 
-@property (nonatomic, assign) int requestTag; /*这个与相应的请求的tag是同步的*/
+@property (nonatomic, assign) NSInteger requestTag; /*这个与相应的请求的tag是同步的*/
 @property (retain, nonatomic) NSString *errorMessage;
