+++ MySuperApp/Class/Databsae/other/EADataModel/提交订单/SubmitOrderSubmitOrderModel.h
@@ -22,7 +22,7 @@
 
 
 @property (retain, nonatomic) NSString *errorMessage;
-@property (nonatomic, assign) int requestTag; /*这个与相应的请求的tag是同步的*/
+@property (nonatomic, assign) NSInteger requestTag; /*这个与相应的请求的tag是同步的*/
 
 + (SubmitOrderSubmitOrderModel *)modelObjectWithDictionary:(NSDictionary *)dict;
 - (id)initWithDictionary:(NSDictionary *)dict;
