+++ MySuperApp/Class/Other/Additions/UIImage+Resize.m
@@ -105,7 +105,7 @@ static CGRect swapWidthAndHeight(CGRect rect)
             break;
             
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", contentMode];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %ld", (long)contentMode];
     }
     
     CGSize newSize = CGSizeMake(self.size.width * ratio, self.size.height * ratio);
