+++ MySuperApp/Class/Other/Toolbar/RKTabView.h
@@ -19,9 +19,9 @@ static inline HorizontalEdgeInsets HorizontalEdgeInsetsMake (CGFloat left, CGFlo
 @protocol RKTabViewDelegate <NSObject>
 
 //Called for all types except TabTypeButton
-- (void)tabView:(RKTabView *)tabView tabBecameEnabledAtIndex:(int)index tab:(RKTabItem *)tabItem;
+- (void)tabView:(RKTabView *)tabView tabBecameEnabledAtIndex:(NSInteger)index tab:(RKTabItem *)tabItem;
 //Called Only for unexcludable items. (TabTypeUnexcludable)
-- (void)tabView:(RKTabView *)tabView tabBecameDisabledAtIndex:(int)index tab:(RKTabItem *)tabItem;
+- (void)tabView:(RKTabView *)tabView tabBecameDisabledAtIndex:(NSInteger)index tab:(RKTabItem *)tabItem;
 
 @end
