+++ MySuperApp/Class/Other/Toolbar/RKTabView.m
@@ -156,7 +156,7 @@
 #pragma mark - Helper methods
 
 - (UIControl *)existingTabForTabItem:(RKTabItem *)tabItem {
-    int index = [self indexOfTab:tabItem];
+    NSInteger index = [self indexOfTab:tabItem];
     if (index != NSNotFound && self.tabViews.count > index) {
         return self.tabViews[[self indexOfTab:tabItem]];
     } else {
@@ -173,7 +173,7 @@
     return self.frame.size.height;
 }
 
-- (int)indexOfTab:(RKTabItem *)tabItem {
+- (NSInteger)indexOfTab:(RKTabItem *)tabItem {
     return [self.tabItems indexOfObject:tabItem];
 }
